requirements.txt
pillow
pyyaml
torch
torchvision
requests
aiohttp
tqdm
